Attempt a workaround for Argparse set_defaults quirks
In various versions of Argparse, set_defaults on subparsers either does
or doesn't override defaults of the same name on parent parsers.

Thus, we set nothing on the parent parser so as to avoid overriding,
then just check whether the attribute exists on the namespace when
attempting to see which command should be ran.